Move digest file to src to get it out of the way of the olm / e2e stuff
