+++ src/digest.ts
